requirements-clip.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
open-clip-torch==2.24.0
torch>=2.2.0
numpy>=1.26.0
Pillow>=10.3.0
python-multipart>=0.0.9
